requirements.txt
gradio==4.44
langchain==1.2.8
langchain-community==0.4.1
llama-index
llama-index-core
fastapi==0.112.4
transformers==4.57.6
pynvml==12.0.0
datasets==4.0.0
openpyxl==3.1.5
trl==0.25.1
peft==0.17.1
bitsandbytes==0.49.1
docx2txt==0.9
torch
torchvision
torchaudio
jupyter==1.1.1
langchainhub==0.1.21
sentence-transformers==5.1.1
faiss-gpu-cu12==1.13.0
accelerate==1.10.1
ninja==1.13.0
wandb==0.22.2
docx==0.2.4
chromadb==1.4.1
pypdf==6.4.0
ragas==0.3.9
jedi==0.19.2
langchain-huggingface-1.2.0
langchain-core==1.2.9